changes in EL Layer

+++ Src/Trials/ConsoleApp/EL/LUT_CIF_TYPE.cs
@@ -5,7 +5,7 @@ namespace SIS.OpenCore.EL
 {
     public partial class LUT_CIF_TYPE
     {
-        public short Code { get; set; }
+        public byte Code { get; set; }
         public string Name { get; set; }
     }
 }
